Replace speedhack DLL and use double speed value
DLL includes fixes and changes speed value from float to double which produces more stable results

+++ DS2 META/TabControls/PlayerControl.xaml.cs
@@ -266,7 +266,7 @@ namespace DS2_META
         private void nudSpeed_ValueChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
         {
             if (GameLoaded && Hook.Hooked)
-                Hook.SetSpeed((float)nudSpeed.Value);
+                Hook.SetSpeed((double)nudSpeed.Value);
         }
         private void nudSpeed_LostFocus(object sender, RoutedEventArgs e)
         {
